+++ simfin/__init__.py
@@ -1,5 +1,5 @@
 # This is also defined in setup.py and must be updated in both places.
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 
 # Expose the following as top-level imports.
 
@@ -12,6 +12,7 @@ from simfin.bulk import load_cashflow, load_cashflow_banks, load_cashflow_insura
 from simfin.utils import add_date_offset
 from simfin.resample import resample, resample_daily, asfreq, asfreq_daily
 from simfin.resample import reindex, index_union
+from simfin.rel_change import rel_change
 
 from simfin.config import set_api_key, load_api_key, get_api_key
 from simfin.config import set_data_dir, get_data_dir, get_download_dir